apply convention plugin only for internal development
