+++ build-logic-internal/gradle.properties
