+++ build-logic/src/main/kotlin/DummyPlugin.kt
@@ -1,0 +1,6 @@
+import org.gradle.api.Plugin
+import org.gradle.api.Project
+
+class DummyPlugin : Plugin<Project> {
+    override fun apply(target: Project) {}
+}
